Plugin and theme updates
This updates plugins and themes to keep this application in sync
with the changes happening in the production network.

The following updates needed to be noted in composer.json due
to version constraints:

* Classic Widgets
* Gutenberg
* Jetpack
* Media Library Assistant
* Shortcode Variables
* WP Sentry Integration

Another change to composer.json is the addition of an allowed
plugins block. This was a prompted change during the update
process.

The full list of plugins affected is

* Addthis
* Akismet
* Black Studio TinyMCE Widget
* CF7 Conditional Fields
* CF7 to Zapier
* Classic Widgets
* CMS Tree Page View
* Contact Form 7
* Custom Post Type UI
* Enable Media Replace
* Google Sitemap Generator
* Gutenberg
* Insert HTML Snippet
* Jetpack
* Media Library Assistant
* Notification
* Page Links To
* Post Snippets
* Redirection
* Shortcode Variables
* Ultimate Posts Widget
* W3 Total Cache
* WP Security Audit Log
* WP Sentry Integration
* WP CF7 Recaptcha

Removed plugins

* Img Title Removal
* Improved Include Page

Some themes were also updated

* TwentyNineteen
* TwentyTwenty
* TwentyTwentyOne